+++ DarkOpt/__init__.py
@@ -1,7 +1,4 @@
-from . import *
-from ._absorber import *
-from ._TES import *
-from ._QET import *
-from ._detector import *
-from ._MaterialProperties import *
-from ._optimize import *
+from . import core
+from .core import *
+from . import materials
+from . import optimize
