+++ DarkOpt/core/_detector.py
@@ -1,7 +1,5 @@
-from darkopt._TES import TES
-from darkopt._QET import QET
-from darkopt._absorber import Absorber
-from darkopt._MaterialProperties import TESMaterial, DetectorMaterial, QETMaterial
+from darkopt.core import TES, QET, Absorber
+from darkopt.materials._MaterialProperties import TESMaterial, DetectorMaterial, QETMaterial
 import numpy as np
 import scipy.constants as constants
 import qetpy as qp
