+++ DarkOpt/materials/_MaterialProperties.py
